refactor(infrastructure): split HydrologyLayer into focused services
Refactor 706-line HydrologyLayer monolith into 6 specialized services
following the established pattern (67% size reduction: 706 → 228 lines).

Services extracted:
- FlowCalculationService (153 LOC): D8 flow directions and accumulation
- StreamCalculationService (157 LOC): Stream identification and Strahler ordering
- MoistureCalculationService (116 LOC): Moisture levels from water/geology
- WaterDepthCalculationService (103 LOC): Stream and pool depth calculation
- SegmentGenerationService (91 LOC): Stream polylines for visualization
- SpringGenerationService (65 LOC): Spring placement at geological boundaries

HydrologyLayer refactored:
- Now lightweight orchestrator (228 lines)
- Injects 6 services via constructor
- Delegates to services in generation pipeline order
- Maintains identical public API and deterministic behavior

DI configuration:
- Added hydrology services to infrastructure.module.ts
- Updated layers/index.ts to export hydrology/*
- All services use @Injectable() with @Optional() @Inject('ILogger')

Pattern: [Concept][Verb]Service
- 4 Calculation services (Flow, Stream, WaterDepth, Moisture)
- 2 Generation services (Spring, Segment)

Benefits:
- Single Responsibility Principle (each service ~65-157 LOC)
- Easier testing (services can be tested independently)
- Better maintainability (focused concerns)
- Follows CLAUDE.md guidelines (no files > 100 lines with multiple concepts)
