+++ packages/infrastructure/src/map/services/layers/index.ts
@@ -16,6 +16,8 @@ export * from './topography';
 
 // Hydrological flow layer implementation
 export { HydrologyLayer } from './HydrologyLayer';
+// Hydrology internal services
+export * from './hydrology';
 
 // Vegetation growth layer implementation
 export { VegetationLayer } from './VegetationLayer';
